Fix API-Gateway. Fix OpenFeign client dynamic services discovery.

+++ seeyousun-backend/review-service/src/main/java/com/taass/seeyousun/reviewservice/repository/ReviewRepository.java
@@ -6,6 +6,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Repository
@@ -13,5 +14,5 @@ public interface ReviewRepository extends JpaRepository<Review, Long> {
     List<Review> findByResortIdOrderByRating(Long resortId);
 
     @Query("SELECT AVG(r.rating) FROM Review r WHERE r.resortId = :resortId")
-    Double findAverageRatingOfResort(@Param("resortId")Long resortId);
+    BigDecimal findAverageRatingOfResort(@Param("resortId")Long resortId);
 }
